scripts/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.23.0
matplotlib>=3.6.0
requests>=2.28.0
